Extract wheel effect apply form into its own component
WheelView.vue mixed wheel display, effect details, and the apply-form logic in one file; move the apply form into WheelEffectApplyForm.vue to keep the view focused on the wheel and history.
